requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dacite==1.9.2
dataclasses-json==0.6.7
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
gradio==5.39.0
gradio_client==1.11.0
greenlet==3.2.3
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
ImageHash==4.3.1
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.1
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.9
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.3
multidict==6.6.3
multimethod==1.12
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.0.1
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
patsy==1.0.1
phik==0.12.5
pillow==11.3.0
propcache==0.3.2
protobuf==5.29.5
puremagic==1.30
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
ruff==0.12.7
safehttpx==0.1.6
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
statsmodels==0.14.5
streamlit==1.47.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
typeguard==4.4.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
visions==0.8.1
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
yarl==1.20.1
ydata-profiling==4.16.1
zstandard==0.23.0
